requirements.txt
slackclient==v1.3.0
anki_vector==v0.5.1
parameterized==v0.6.1
Pillow>=3.3